reDreamIRT/pyproject.toml
[project]
name = "redreamai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi~=0.111.1",
    "uvicorn~=0.30.3",
    "python-dotenv~=1.0.1",
    "openai~=1.75.0",
    "groq~=0.23.0",
    "langfuse~=2.60.3",
    "tiktoken~=0.9.0",
    "pydantic>=2.11.3",
    "pytest>=8.3.5",
    "aiohttp"
]
